requirements.txt
# AWS Integration
boto3==1.34.19
botocore==1.34.19

# CLI Framework
click==8.1.7
rich==13.7.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data Analysis
pandas==2.1.4
numpy==1.26.2
papermill==2.4.0
jupyter==1.0.0
nbconvert==7.12.0
mistune==2.0.5
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
perfetto==0.16.0
beautifulsoup4==4.12.3

# Git Operations
gitpython==3.1.40

# HTTP/Download
requests==2.31.0

# Validation
pydantic==2.5.3
typing-extensions==4.9.0

# Development/Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.7.1
